refactor: strategies  폴더에 strategy 파일 이동
