+++ src/auth/strategies/local.strategy.ts
@@ -1,8 +1,8 @@
 import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Strategy } from 'passport-local';
-import { AuthService } from './auth.service';
-import { AuthInfo } from './types/auth-info';
+import { AuthService } from '../auth.service';
+import { AuthInfo } from '../types/auth-info';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
